+++ lib/movies/data/models/top_rated_movies_model.dart
@@ -9,7 +9,7 @@ class TopRatedMoviesModel extends TopRatedEntities{
       genreIds: List<int>.from(json["genre_ids"].map((e) => e)),
       overview: json["overview"],
       title: json["title"],
-      voteAverage: json["vote_average"].toDouble,
+      voteAverage: json["vote_average"],
       releaseDate: json["release_date"],
     );
   }
